move common helper in polyglot-po
