--- packages/common/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "common",
-  "private": true,
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {},
-  "dependencies": {
-    "lodash": "4.17.15",
-    "polyglot-po": "^0.0.1"
-  },
-  "author": "",
-  "license": "Apache-2.0"
-}
