+++ packages/polyglot-po/src/dictionaryToObject.js
