+++ packages/polyglot-po/src/extractPoEntry.js
